CPU TIme solution
It now logs the total cpu time on a specific action by threading the operations, plus we avoid blocking the node for other actions. A primary code has been proposed for usability notifications, have an idea what the node is doing, we can also add a tv below the details saying what the node is doing, updating it periodically.